pyproject.toml
[tool.poetry]
name = "pre-commit-mypy-poetry-sync-hook"
version = "0.1.0"
description = "A pre-commit hook to keep mypy pre-commit hook's additional_dependencies in sync with poetry"
authors = ["Martin Schulze <martin.schulze@e2m.de>"]
license = "MIT"
readme = "README.md"
packages = [{include="pre_commit_mypy_poetry_sync_hook/*.py"}]

[tool.poetry.scripts]
sync-mypy-additional-dependencies = "pre_commit_mypy_poetry_sync_hook.sync_mypy_additional_dependencies:main"

[tool.poetry.dependencies]
python = "^3.10"
ruamel-yaml = "^0.18.10"
poetry-core = "^2.0.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"